+++ cmd/mq_prometheus/mq_prometheus.sh
@@ -26,4 +26,4 @@ channels="TO.*,SYSTEM.DEF.SVRCONN"
 # Start via "exec" so the pid remains the same. The queue manager can
 # then check the existence of the service and use the MQ_SERVER_PID value
 # to kill it on shutdown.
-exec /usr/local/bin/mqgo/mq_prometheus -ibmmq.queueManager=$qMgr -ibmmq.monitoredQueues="$queues" -ibmmq.monitoredChannels="$channels" -log.level=error -ibmmq.qStatus=true
+exec /usr/local/bin/mqgo/mq_prometheus -ibmmq.queueManager=$qMgr -ibmmq.monitoredQueues="$queues" -ibmmq.monitoredTopics="#" -ibmmq.monitoredChannels="$channels" -log.level=error -ibmmq.qStatus=true