+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 Newest updates are at the top of this file.
 
+## April 2019
+* Add topic and queue manager status support from latest mqmetric library
+* Add Grafana/Prometheus dashboards showing the newer metrics
+* Update InfluxDB collector to similar level as Prometheus/JSON
+
 ## March 2019
 * Update to use v4.0.1 of the mq-golang repository
 * Update READMEs for all the monitor agent programs